북, 페이지, 공지사항 메인페이지 기능 완성

+++ finalProject_maper/src/main/java/com/mapers/book/service/BookDeleteController.java
@@ -13,25 +13,28 @@ public class BookDeleteController implements Controller {
 	@Override
 	public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
 		System.out.println("북 딜리트 컨트롤러 호출");
-		
-		BookDTO bookDTO = (BookDTO)request.getSession().getAttribute("bookDTO");
-		
-		int idx = bookDTO.getBookNum();
-		System.out.println(idx);
 
-		BookDAO dao = BookDAO.getInstance();
+		BookDTO bookDTO = (BookDTO) request.getSession().getAttribute("bookDTO");
+		if (bookDTO == null) {
+			return "/Common/logOutProcess.jsp";
+		} else {
 
-		BookDTO dto = dao.selectBook(idx);
-		System.out.println("셀렉트 북 성공");
-		int result = dao.deleteBook(idx);
-		if (result == 1) {
-			String saveFileName = dto.getSfile();
-			FileUtil.deleteFile(request, "/Uploads/Book", saveFileName);
+			int idx = bookDTO.getBookNum();
+
+			BookDAO dao = BookDAO.getInstance();
+
+			BookDTO dto = dao.selectBook(idx);
+			System.out.println("셀렉트 북 성공");
+			dao.deleteAllLikeList(idx);
+			int result = dao.deleteBook(idx);
+			if (result == 1) {
+				String saveFileName = dto.getSfile();
+				FileUtil.deleteFile(request, "/Uploads/Book", saveFileName);
+			}
+			System.out.println("북 삭제 성공");
+
+			request.setAttribute("url", "/Book/book.do?command=bookList");
+			return "redirect:../Book/book.do?command=bookList";
 		}
-		System.out.println("북 삭제 성공");
-		dao.close();
-
-		request.setAttribute("url", "/Book/book.do?command=bookList");
-		return "redirect:../Book/book.do?command=bookList";
 	}
 }
